% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_block_number.R
\name{eth_block_number}
\alias{eth_block_number}
\title{Returns the latest block number of the blockchain.}
\usage{
eth_block_number(provider_url, api_key)
}
\arguments{
\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns the latest block number of the blockchain.
}
\description{
Returns the latest block number of the blockchain.
}
\examples{
my_provider_url = "https://eth-mainnet.rpc.grove.city/v1/"
my_api_key = "fadc98f89"

eth_block_number(my_provider_url, my_api_key)

}
