% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_chain_id.R
\name{eth_chain_id}
\alias{eth_chain_id}
\title{Returns the currently configured chain id}
\usage{
eth_chain_id(provider_url, api_key)
}
\arguments{
\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns the currently configured chain id
}
\description{
Returns the currently configured chain id
}
\examples{
my_provider_url = "https://eth-mainnet.rpc.grove.city/v1/"
my_api_key = "fadc98f89"

eth_chain_id(my_provider_url, my_api_key)

}
