% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_estimateGas.R
\name{eth_estimateGas}
\alias{eth_estimateGas}
\title{Fill in later}
\usage{
eth_estimateGas(sender, receiver, provider_url, api_key)
}
\arguments{
\item{sender}{An Ethereum wallet address (ENS domains not supported).}

\item{receiver}{An Ethereum wallet address (ENS domains not supported).}

\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
A numeric value of an Ethereum balance.
}
\description{
Fill in later
}
\examples{
vitalik_eth = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045"
hayden_eth = "0x50EC05ADe8280758E2077fcBC08D878D4aef79C3"
my_provider_url = "https://eth-mainnet.rpc.grove.city/v1/"
my_api_key = "fadc98f89"

eth_estimateGas(sender = vitalik_eth, receiver = hayden_eth, 
                 provider_url = my_provider_url, api_key = my_api_key)

}
