% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_feeHistory.R
\name{eth_feeHistory}
\alias{eth_feeHistory}
\title{Returns the collection of historical gas information.}
\usage{
eth_feeHistory(
  block_count,
  newest_block,
  reward_min,
  reward_max,
  provider_url,
  api_key
)
}
\arguments{
\item{block_count}{The number of blocks in the requested range. Between 1 and 1024 blocks can be requested in a single query. It will return less than the requested range if not all blocks are available.}

\item{newest_block}{The highest number block of the requested range in hexadecimal format or tags. The supported tag values include earliest for the earliest/genesis block, latest for the latest mined block, pending for the pending state/transactions, safe for the most recent secure block, and finalized for the most recent secure block accepted by more than 2/3 of validators.}

\item{reward_min}{The minimum percentile value that surrounds the range of percentile rewards}

\item{reward_max}{The maximum percentile value that surrounds the range of percentile rewards}

\item{provider_url}{The URL of a JSON-RPC provider such as Grove, DRPC, Quicknode, Infura, Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns the collection of historical gas information: oldestBlock, baseFeePerGas, gasUsedRatio and reward.
}
\description{
Returns the collection of historical gas information.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"

eth_feeHistory(block_count = 4, newest_block = "latest",
               reward_min = 25, reward_max = 75,
               my_provider_url, my_api_key)

}
