% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_gasPrice.R
\name{eth_gasPrice}
\alias{eth_gasPrice}
\title{Returns the current gas price in ETH, gwei, or wei.}
\usage{
eth_gasPrice(provider_url, api_key)
}
\arguments{
\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
A numeric value of an Ethereum balance.
}
\description{
Returns the current gas price in ETH, gwei, or wei.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"

eth_gasPrice(my_provider_url, my_api_key)

}
