% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_get_balance.R
\name{eth_get_balance}
\alias{eth_get_balance}
\title{Get an Ethereum wallet balance}
\usage{
eth_get_balance(wallet_address, provider_url, api_key)
}
\arguments{
\item{wallet_address}{An Ethereum wallet balance (ENS domains not supported).}

\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Grove, or Infura.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
A numeric value of an Ethereum balance.
}
\description{
Get an Ethereum wallet balance
}
\examples{
vitalik_eth = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045"
my_provider_url = "https://eth-mainnet.rpc.grove.city/v1/"
my_api_key = "fadc98f89"

eth_get_balance(vitalik_eth, my_provider_url, my_api_key)

}
