% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_getBalance.R
\name{eth_getBalance}
\alias{eth_getBalance}
\title{Get the balance of the specified address in eth, wei, or hexidecimal format}
\usage{
eth_getBalance(wallet_address, provider_url, api_key, return_format)
}
\arguments{
\item{wallet_address}{An Ethereum wallet balance (ENS domains not supported).}

\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}

\item{return_format}{Returns the balance of wallet in "eth", "wei", or "hexidecimal" format}
}
\value{
A numeric value of an Ethereum balance.
}
\description{
Get the balance of the specified address in eth, wei, or hexidecimal format
}
\examples{
vitalik_eth = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045"
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"
return_format = "ETH"

eth_getBalance(vitalik_eth, my_provider_url, my_api_key, return_format)

}
