% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_getBlockByHash.R
\name{eth_getBlockByHash}
\alias{eth_getBlockByHash}
\title{Returns information of the block matching the given block hash.}
\usage{
eth_getBlockByHash(block_hash, provider_url, api_key)
}
\arguments{
\item{block_hash}{The hash (32 bytes) of the block}

\item{provider_url}{The URL of a JSON-RPC provider such as Grove, DRPC, Quicknode, Infura, Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns information of the block matching the given block hash.
}
\description{
Returns information of the block matching the given block hash.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"
my_block_hash = "0x3f07a9c83155594c000642e7d60e8a8a00038d03e9849171a05ed0e2d47acbb3"

eth_getBlockByHash(my_block_hash, my_provider_url, my_api_key)

}
