% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_getBlockByNumber.R
\name{eth_getBlockByNumber}
\alias{eth_getBlockByNumber}
\title{Returns information of the block matching the given block number.}
\usage{
eth_getBlockByNumber(block_number, provider_url, api_key)
}
\arguments{
\item{block_number}{The block number in hexadecimal format or the string "earliest", "latest", "pending", "safe", or "finalized".}

\item{provider_url}{The URL of a JSON-RPC provider such as Grove, DRPC, Quicknode, Infura, Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns information of the block matching the given block number.
}
\description{
Returns information of the block matching the given block number.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"

eth_getBlockByNumber("0xc5043f", my_provider_url, my_api_key)

}
