% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_getBlockTransactionCountByHash.R
\name{eth_getBlockTransactionCountByHash}
\alias{eth_getBlockTransactionCountByHash}
\title{Returns the number of transactions for the block matching the given block hash.}
\usage{
eth_getBlockTransactionCountByHash(block_hash, provider_url, api_key)
}
\arguments{
\item{block_hash}{A block hash}

\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns the number of transactions for the block matching the given block hash.
}
\description{
Returns the number of transactions for the block matching the given block hash.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"
my_block_hash = "0x829df9bb801fc0494abf2f443423a49ffa32964554db71b098d332d87b70a48b"

eth_getBlockTransactionCountByHash(my_block_hash, my_provider_url, my_api_key)

}
