% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_getBlockTransactionCountByNumber.R
\name{eth_getBlockTransactionCountByNumber}
\alias{eth_getBlockTransactionCountByNumber}
\title{Returns the number of transactions for the block matching the given block number.}
\usage{
eth_getBlockTransactionCountByNumber(block_number, provider_url, api_key)
}
\arguments{
\item{block_number}{A block number as a string in hexadecimal format or tags.}

\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns the number of transactions for the block matching the given block number.
}
\description{
Returns the number of transactions for the block matching the given block number.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"
my_block_number = "0xc5043f"

eth_getBlockTransactionCountByNumber(my_block_number, my_provider_url, my_api_key)

}
