% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_getCode.R
\name{eth_getCode}
\alias{eth_getCode}
\title{Returns the compiled byte code of a smart contract, if any, at a given address.}
\usage{
eth_getCode(wallet_address, provider_url, api_key)
}
\arguments{
\item{wallet_address}{An Ethereum wallet balance (ENS domains not supported).}

\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns the compiled byte code of a smart contract, if any, at a given address.
}
\description{
Returns the compiled byte code of a smart contract, if any, at a given address.
}
\examples{
vitalik_eth = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045"
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"

eth_getCode(vitalik_eth, my_provider_url, my_api_key)

}
