% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_getTransactionByBlockHashAndIndex.R
\name{eth_getTransactionByBlockHashAndIndex}
\alias{eth_getTransactionByBlockHashAndIndex}
\title{Returns information about a transaction given a blockhash and transaction index position.}
\usage{
eth_getTransactionByBlockHashAndIndex(block_hash, index, provider_url, api_key)
}
\arguments{
\item{block_hash}{A block hash.}

\item{index}{An integer of the transaction index position encoded as a hexadecimal.}

\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns information about a transaction given a blockhash and transaction index position.
}
\description{
Returns information about a transaction given a blockhash and transaction index position.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"
my_block_hash = "0x829df9bb801fc0494abf2f443423a49ffa32964554db71b098d332d87b70a48b"
my_index = "0x0" 

eth_getTransactionByBlockHashAndIndex(my_block_hash, my_index, 
                                      my_provider_url, my_api_key)

}
