% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_getTransactionByBlockNumberAndIndex.R
\name{eth_getTransactionByBlockNumberAndIndex}
\alias{eth_getTransactionByBlockNumberAndIndex}
\title{Returns information about a transaction given a block number and transaction index position.}
\usage{
eth_getTransactionByBlockNumberAndIndex(
  block_number,
  index,
  provider_url,
  api_key
)
}
\arguments{
\item{block_number}{A block number as a string in hexadecimal format or tags.}

\item{index}{An integer of the transaction index position encoded as a hexadecimal.}

\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns information about a transaction given a block number and transaction index position.
}
\description{
Returns information about a transaction given a block number and transaction index position.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"
my_block_number = "0xc5043f"
my_index = "0x0" 

eth_getTransactionByBlockNumberAndIndex(my_block_number, my_index, 
                                        my_provider_url, my_api_key)

}
