% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_getTransactionCount.R
\name{eth_getTransactionCount}
\alias{eth_getTransactionCount}
\title{Returns the number of transactions sent from an address.}
\usage{
eth_getTransactionCount(
  address,
  block_number = c("earliest", "latest", "pending", "safe", "finalized"),
  provider_url,
  api_key
)
}
\arguments{
\item{address}{The address from which the transaction count to be checked}

\item{block_number}{The block number as a string in hexadecimal format or tags. Supported tags include "earliest", "latest", "pending", "safe" and "finalized".}

\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns the number of transactions sent from an address.
}
\description{
Returns the number of transactions sent from an address.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"

eth_accounts(my_provider_url, my_api_key)

}
