% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_getTransactionReceipt.R
\name{eth_getTransactionReceipt}
\alias{eth_getTransactionReceipt}
\title{Returns the receipt of a transaction by transaction hash.}
\usage{
eth_getTransactionReceipt(block_hash, provider_url, api_key)
}
\arguments{
\item{block_hash}{A transaction hash}

\item{provider_url}{The URL of a JSON-RPC provider such as Grove, DRPC, Quicknode, Infura, Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns the receipt of a transaction by transaction hash.
}
\description{
Returns the receipt of a transaction by transaction hash.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"
my_block_hash = "0x85d995eba9763907fdf35cd2034144dd9d53ce32cbec21349d4b12823c6860c5"

eth_getTransactionReceipt(my_block_hash, my_provider_url, my_api_key)

}
