% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_protocolVersion.R
\name{eth_protocolVersion}
\alias{eth_protocolVersion}
\title{Returns the current Ethereum protocol version}
\usage{
eth_protocolVersion(provider_url, api_key)
}
\arguments{
\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns the current Ethereum protocol version
}
\description{
Returns the current Ethereum protocol version
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"

eth_protocolVersion(my_provider_url, my_api_key)

}
