% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorerBalance.R
\name{explorerBalance}
\alias{explorerBalance}
\title{Get the balance of the specified address in ETH}
\usage{
explorerBalance(wallet_address, blockchain_explorer, api_key)
}
\arguments{
\item{wallet_address}{An Ethereum wallet balance (ENS domains not supported).}

\item{blockchain_explorer}{A blockchain explorer API provider. Supported}

\item{api_key}{The API key of the Blockchain Explorer such as Etherscan or Arbiscan.}
}
\value{
Returns a dataframe with columns of account address, account balance, and currency.
}
\description{
Get the balance of the specified address in ETH
}
\examples{

explorerBalance(wallet_address = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045",
                blockchain_explorer = "etherscan",     
                api_key = "A79S4QJXS518CS6ZQYFHB5SX5AWE7E952D")

}
