% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorerBalances.R
\name{explorerBalances}
\alias{explorerBalances}
\title{Get the balance of multiple addresses in ETH}
\usage{
explorerBalances(wallet_addresses, blockchain_explorer, api_key)
}
\arguments{
\item{wallet_addresses}{An Ethereum wallet balance (ENS domains not supported).}

\item{blockchain_explorer}{A blockchain explorer API provider: "etherscan", "arbiscan".}

\item{api_key}{The API key of the Blockchain Explorer such as Etherscan or Arbiscan.}
}
\value{
Returns a dataframe with columns of account addresses, account balances, and currencies.
}
\description{
Get the balance of multiple addresses in ETH
}
\examples{
blockchain_wallets = c("0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045",
                       "0x50EC05ADe8280758E2077fcBC08D878D4aef79C3")

explorerBalances(wallet_addresses = blockchain_wallets,
                 blockchain_explorer = "ethereum",     
                 api_key = "A79S4QJXS518CS6ZQYFHB5SX5AWE7E952D")

}
