% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCryptoPrice.R
\name{getCryptoPrice}
\alias{getCryptoPrice}
\title{Returns an array of addresses owned by the client. Note that most providers
do not store private keys, so this will often return an empty response.}
\usage{
getCryptoPrice(token, api_key)
}
\arguments{
\item{token}{The token that you want the price of relative to USD.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns an array of addresses owned by the client.
}
\description{
Returns an array of addresses owned by the client. Note that most providers
do not store private keys, so this will often return an empty response.
}
\examples{
my_api_key = "3a2f2d6ce1672801c3a044139373a287"

getCryptoPrice("GRT", my_api_key)

}
