% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExplorerURL.R
\name{getExplorerURL}
\alias{getExplorerURL}
\title{Returns a link to a blockchain explorer}
\usage{
getExplorerURL(blockchain_explorer)
}
\arguments{
\item{blockchain_explorer}{A blockchain explorer API provider: "etherscan", "arbiscan".}
}
\value{
Returns a link to a blockchain explorer - typically used in explorerBalances().
}
\description{
Returns a link to a blockchain explorer
}
\examples{

getExplorerURL("etherscan")

}
