% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_listening.R
\name{net_listening}
\alias{net_listening}
\title{Returns \code{true} if client is actively listening for network connections}
\usage{
net_listening(provider_url, api_key)
}
\arguments{
\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns \code{true} if client is actively listening for network connections
}
\description{
Returns \code{true} if client is actively listening for network connections
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"

net_listening(my_provider_url, my_api_key)

}
