% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_version.R
\name{net_version}
\alias{net_version}
\title{Returns the current network ID.}
\usage{
net_version(provider_url, api_key)
}
\arguments{
\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns the current network ID.
}
\description{
Returns the current network ID.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"

net_version(my_provider_url, my_api_key)

}
