% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_hosted_subgraph.R
\name{query_hosted_subgraph}
\alias{query_hosted_subgraph}
\title{Returns subgraph data from an indexed smart contract on The Graph's hosted service}
\usage{
query_hosted_subgraph(graphql_query, subgraph_url)
}
\arguments{
\item{graphql_query}{A GraphQL query of a subgraph.}

\item{subgraph_url}{The URL of a hosted service subgraph from The Graph.}
}
\value{
Returns subgraph data from an indexed smart contract.
}
\description{
Returns subgraph data from an indexed smart contract on The Graph's hosted service
}
\examples{
uniswap_query = "{factories(first: 5){ id poolCount txCount totalVolumeUSD }}"
uniswap_v3_hosted = "https://api.thegraph.com/subgraphs/name/uniswap/uniswap-v3"

query_hosted_subgraph(uniswap_query, uniswap_v3_hosted)

}
