% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web3_clientVersion.R
\name{web3_clientVersion}
\alias{web3_clientVersion}
\title{Returns the current client version.}
\usage{
web3_clientVersion(provider_url, api_key)
}
\arguments{
\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns the current client version.
}
\description{
Returns the current client version.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"

web3_clientVersion(my_provider_url, my_api_key)

}
