% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web3_sha3.R
\name{web3_sha3}
\alias{web3_sha3}
\title{Returns Keccak-256 (not the standardized SHA3-256) hash of the given data}
\usage{
web3_sha3(sha3, provider_url, api_key)
}
\arguments{
\item{sha3}{The data in hexadecimal form to convert into a SHA3 hash}

\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
The SHA3 hash of the given string
}
\description{
Returns Keccak-256 (not the standardized SHA3-256) hash of the given data
}
\examples{
my_sha3 = "0x68656c6c6f20776f726c64"
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"

web3_sha3(my_sha3, my_provider_url, my_api_key)

}
