% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eth_accounts.R
\name{eth_accounts}
\alias{eth_accounts}
\title{Returns an array of addresses owned by the client. Note that most providers
do not store private keys, so this will often return an empty response.}
\usage{
eth_accounts(provider_url, api_key)
}
\arguments{
\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns an array of addresses owned by the client.
}
\description{
Returns an array of addresses owned by the client. Note that most providers
do not store private keys, so this will often return an empty response.
}
\examples{
my_provider_url = "https://mainnet.infura.io/v3/"
my_api_key = "04f25f82de3840e1afd524b62d6767c5"

eth_accounts(my_provider_url, my_api_key)

}
