% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_listening.R
\name{net_listening}
\alias{net_listening}
\title{Returns \code{true} if client is actively listening for network connections}
\usage{
net_listening(provider_url, api_key)
}
\arguments{
\item{provider_url}{The URL of a JSON-RPC provider such as Pocket Network, Infura, or Alchemy.}

\item{api_key}{The API key of the JSON-RPC provider.}
}
\value{
Returns \code{true} if client is actively listening for network connections
}
\description{
Returns \code{true} if client is actively listening for network connections
}
\examples{
my_provider_url = "https://eth-mainnet.rpc.grove.city/v1/"
my_api_key = "fadc98f89"

net_listening(my_provider_url, my_api_key)

}
